Tests/LibWeb/Text/expected/css/input-line-height-clamp.txt

small: normal
large: 40px
variable: normal
checkbox: 1px
